refactoring se
